refactor(data-model): move the JSON primitive codecs into `codec-json`
`BigIntCodec`, `SpecialNumberCodec`, `SymbolCodec` and `UndefinedCodec` each
exist because JSON cannot carry the type it handles, and their only registrant
is `createBaseJsonRegistry()`. That makes them part of the format rather than
part of the shared machinery, so they move to `codec-json/` with their tests.

The payoff is larger than the tidiness, and it settles something the earlier
split could not. Three of the four imported `ProblematicValue`, and they were
the last references to a fabric class anywhere in `codec-common/`. With them
gone, that directory reaches the value domain only through `import type`, so
it has no runtime dependency on a fabric class at all -- it is now purely the
registry, the codec and reconstruction-context base classes, and the tag
vocabularies. Its barrel says so, since that is the invariant a later
convenience import would quietly undo.

`SymbolCodec` goes with the others even though `structuredClone` rejects a
symbol where it carries `bigint`, `undefined` and the special numbers
natively, so a second format would need symbol handling of its own. Dan's
call: that handling is expected to differ enough that sharing this file would
not pay, and leaving it behind would forfeit the whole `codec-common` result
for one file.

`main` is red independently of this change -- a semantic conflict between two
merges, in `packages/cli`. Verified identical before and after: the same nine
`TS2561` errors at the same nine locations in `piece-call.test.ts`, and the
same two `cli` test failures. The workspace suite was run with `cli` excluded
for that reason; everything else passes.
