+++ packages/data-model/src/codec-json/SpecialNumberCodec.ts
